// app/api/lookup/guide-language-price/interface/guide-language-price.ts
export interface GuideLanguagePriceEntity {
  is_used?: boolean;
  is_deleted?: boolean;
  is_active: boolean;
  created_at?: string;
  updated_at?: string;
  created_by?: string;
  updated_by?: string;
  host_ip?: string;
  url?: string;
  deleted_at?: string;
  deleted_by?: string;
  tenant_id?: string;
  id: number;
  guide_id: number;
  language_id: number;
  price: string;
}